+++ nengo/tests/test_config.py
@@ -298,15 +298,7 @@ def test_classparams_str_repr():
 def test_config_repr():
     """tests the repr function in Config class"""
     model = nengo.Network()
-    r = repr(model.config)  # == "<Config(Connection, Ensemble, Node, Probe)>
-    assert (  # Python <= 3.5 has types in any order
-        r.startswith("<Config(")
-        and r.endswith(")>")
-        and "Connection" in r
-        and "Ensemble" in r
-        and "Node" in r
-        and "Probe" in r
-    )
+    assert repr(model.config) == "<Config(Connection, Ensemble, Node, Probe)>"
 
 
 def test_config_exit_errors():